create util func and move App.Locals namespace for thing/[id] routes
