+++ src/pages/sights/[id].ts
@@ -1,23 +1,4 @@
-import type { APIRoute } from "astro";
+import { thingIDRewriteRoute } from "../../server/thingUtils";
 
 export const prerender = false;
-
-export const GET: APIRoute = (context) => {
-    const idMatch = context.params.id?.match(/^\d+/);
-
-    if (idMatch == null) {
-        return new Response(null, { status: 404 });
-    }
-
-    context.locals.sightId = Number.parseInt(idMatch[0], 10);
-
-    return context.rewrite("/sights");
-};
-
-declare global {
-    namespace App {
-        interface Locals {
-            sightId?: number;
-        }
-    }
-}
+export const GET = thingIDRewriteRoute("sights");
