+++ src/server/thingUtils.ts
@@ -1,5 +1,5 @@
 import { LibsqlError } from "@libsql/client";
-import type { APIContext } from "astro";
+import type { APIContext, APIRoute } from "astro";
 import { and, eq } from "drizzle-orm";
 import { jsonError, jsonResponse } from "./responses";
 import { paginationQuery, parseNumberCursor } from "./pagination";
@@ -93,4 +93,32 @@ export function thingColourHandler<T extends { memberColor: string, showColour: 
             showColour: undefined,
         };
     });
+}
+
+export function thingIDRewriteRoute(resourceName: string): APIRoute {
+    const localKey = `${resourceName.slice(0, -1)}Id` as keyof App.Locals;
+    const rewritePath = `/${resourceName}`;
+
+    return (context) => {
+        const idMatch = context.params.id?.match(/^\d+/);
+
+        if (idMatch == null) {
+            return new Response(null, { status: 404 });
+        }
+
+        // If it's not any it fucking says bullshit like 'Number is not assignable to type never' like a dumbass so
+        (context.locals as any)[localKey] = Number.parseInt(idMatch[0], 10);
+
+        return context.rewrite(rewritePath);
+    };
+}
+
+declare global {
+    namespace App {
+        interface Locals {
+            soundId?: number;
+            sightId?: number;
+            motionId?: number;
+        }
+    }
 }
